add export conditions in pages
