+++ src/app/(blog)/blogs/[id]/page.js
@@ -171,4 +171,12 @@ const BlogSingle = ({params}) => {
   );
 };
 
+export async function generateStaticParams() {
+  const blogIds = [1, 2, 3, 4]; // O puedes obtener estos IDs dinámicamente desde una API o base de datos
+
+  return blogIds.map(id => ({
+    id: String(id), // Aquí usamos String(id) para asegurarnos de que sea un string
+  }));
+}
+
 export default BlogSingle;
