+++ src/app/(home)/home-v5/page.js
@@ -194,4 +194,12 @@ const Home_V5 = () => {
   );
 };
 
+export async function generateStaticParams() {
+  const blogIds = [1, 2, 3, 4]; // O puedes obtener estos IDs dinámicamente desde una API o base de datos
+
+  return blogIds.map(id => ({
+    id: String(id), // Aquí usamos String(id) para asegurarnos de que sea un string
+  }));
+}
+
 export default Home_V5;
