+++ src/app/(property)/(agents)/agent-single/[id]/page.js
@@ -170,4 +170,12 @@ const AgentSingle = ({params}) => {
   );
 };
 
+export async function generateStaticParams() {
+  const blogIds = [1, 2, 3, 4]; // O puedes obtener estos IDs dinámicamente desde una API o base de datos
+
+  return blogIds.map(id => ({
+    id: String(id), // Aquí usamos String(id) para asegurarnos de que sea un string
+  }));
+}
+
 export default AgentSingle;
