+++ src/app/(property)/(single-style)/single-v1/[id]/SingleComponent.js
@@ -1,0 +1,276 @@
+"use client";
+import DefaultHeader from "@/components/common/DefaultHeader";
+import Footer from "@/components/common/default-footer";
+import MobileMenu from "@/components/common/mobile-menu";
+import EnergyClass from "@/components/property/property-single-style/common/EnergyClass";
+import FloorPlans from "@/components/property/property-single-style/common/FloorPlans";
+import HomeValueChart from "@/components/property/property-single-style/common/HomeValueChart";
+import InfoWithForm from "@/components/property/property-single-style/common/more-info";
+import NearbySimilarProperty from "@/components/property/property-single-style/common/NearbySimilarProperty";
+import OverView from "@/components/property/property-single-style/common/OverView";
+import PropertyAddress from "@/components/property/property-single-style/common/PropertyAddress";
+import PropertyDetails from "@/components/property/property-single-style/common/PropertyDetails";
+import PropertyFeaturesAminites from "@/components/property/property-single-style/common/PropertyFeaturesAminites";
+import PropertyHeader from "@/components/property/property-single-style/common/PropertyHeader";
+import PropertyNearby from "@/components/property/property-single-style/common/PropertyNearby";
+import PropertyVideo from "@/components/property/property-single-style/common/PropertyVideo";
+import PropertyViews from "@/components/property/property-single-style/common/property-view";
+import ProperytyDescriptions from "@/components/property/property-single-style/common/ProperytyDescriptions";
+import ReviewBoxForm from "@/components/property/property-single-style/common/ReviewBoxForm";
+import VirtualTour360 from "@/components/property/property-single-style/common/VirtualTour360";
+import AllReviews from "@/components/property/property-single-style/common/reviews";
+import ContactWithAgent from "@/components/property/property-single-style/sidebar/ContactWithAgent";
+import ScheduleTour from "@/components/property/property-single-style/sidebar/ScheduleTour";
+import PropertyGallery from "@/components/property/property-single-style/single-v1/PropertyGallery";
+import React, { useEffect, useState } from "react";
+import MortgageCalculator from "@/components/property/property-single-style/common/MortgageCalculator";
+import WalkScore from "@/components/property/property-single-style/common/WalkScore";
+import { getListings } from "@/transform/propertiesTransform";
+import { getGalleryTransform } from "@/transform/galeryTransform";
+import Header from "@/components/home/home-v5/Header";
+
+// export const metadata = {
+//   title: "Cygnus",
+// };
+
+const SingleComponent = ({ params }) => {
+
+  const [listings, setListings] = useState([]);
+  const [gallery, setGallery] = useState([]);
+
+  useEffect(() => {
+    const loadListings = async () => {
+      try {
+        const [items, galleryData] = await Promise.all([
+          getListings(),
+          getGalleryTransform(params.id)
+        ]);
+
+        setListings(items);
+        setGallery(galleryData);
+
+      } catch (error) {
+        console.error("Error al cargar los items:", error);
+      }
+    };
+
+    loadListings();
+  }, []);
+
+
+
+  return (
+    <>
+      {/* Main Header Nav */}
+      <Header />
+      <br/>
+      <br/>
+      <br/>
+      <br/>
+      {/* End Main Header Nav */}
+
+      {/* Mobile Nav  */}
+      <MobileMenu />
+      {/* End Mobile Nav  */}
+
+      {/* Property All Single V1 */}
+      <section className="pt60 pb90 bgc-f7">
+        <div className="container">
+          <div className="row">
+            <PropertyHeader id={params.id} listings= {listings} />
+          </div>
+          {/* End .row */}
+
+          <div className="row mb30 mt30">
+            <PropertyGallery id={params.id} listings= {listings} gallery={gallery} />
+          </div>
+          {/* End .row */}
+
+          <div className="row wrap">
+            <div className="col-lg-8">
+              <div className="ps-widget bgc-white bdrs12 default-box-shadow2 p30 mb30 overflow-hidden position-relative">
+                <h4 className="title fz17 mb30">Cuenta con</h4>
+                <div className="row">
+                  <OverView listings={listings}/>
+                </div>
+              </div>
+              {/* End .ps-widget */}
+
+              <div className="ps-widget bgc-white bdrs12 default-box-shadow2 p30 mb30 overflow-hidden position-relative">
+                <h4 className="title fz17 mb30">Descripción de la propiedad</h4>
+                <ProperytyDescriptions id={params.id} listings={listings}/>
+                {/* End property description */}
+
+                <h4 className="title fz17 mb30 mt50">
+                  Detalles de la propiedad
+                </h4>
+                <div className="row">
+                  <PropertyDetails id={params.id} listings={listings}/>
+                </div>
+              </div>
+              {/* End .ps-widget */}
+
+              <div className="ps-widget bgc-white bdrs12 default-box-shadow2 p30 mb30 overflow-hidden position-relative">
+                <h4 className="title fz17 mb30 mt30">Ubicación</h4>
+                <div className="row">
+                  <PropertyAddress id={params.id} listings={listings}/>
+                </div>
+              </div>
+              {/* End .ps-widget */}
+
+              <div className="ps-widget bgc-white bdrs12 default-box-shadow2 p30 mb30 overflow-hidden position-relative">
+                <h4 className="title fz17 mb30">
+                  Caracteristicas &amp; Comodidades
+                </h4>
+                <div className="row">
+                  <PropertyFeaturesAminites id={params.id} listings={listings}/>
+                </div>
+              </div>
+              {/* End .ps-widget */}
+
+              {/* <div className="ps-widget bgc-white bdrs12 default-box-shadow2 p30 mb30 overflow-hidden position-relative">
+                <h4 className="title fz17 mb30">Energy Class</h4>
+                <div className="row">
+                  <EnergyClass />
+                </div>
+              </div> */}
+              {/* End .ps-widget */}
+
+              {/* <div className="ps-widget bgc-white bdrs12 default-box-shadow2 p30 mb30 overflow-hidden position-relative">
+                <h4 className="title fz17 mb30">Floor Plans</h4>
+                <div className="row">
+                  <div className="col-md-12">
+                    <div className="accordion-style1 style2">
+                      <FloorPlans />
+                    </div>
+                  </div>
+                </div>
+              </div> */}
+              {/* End .ps-widget */}
+
+              <div className="ps-widget bgc-white bdrs12 default-box-shadow2 p30 mb30 ">
+                <h4 className="title fz17 mb30">Video</h4>
+                <div className="row">
+                  <PropertyVideo id={params.id} listings={listings}/>
+                </div>
+              </div>
+              {/* End .ps-widget */}
+
+              {/* <div className="ps-widget bgc-white bdrs12 default-box-shadow2 p30 mb30 overflow-hidden position-relative">
+                <h4 className="title fz17 mb30">360° Virtual Tour</h4>
+                <div className="row">
+                  <VirtualTour360 />
+                </div>
+              </div> */}
+              {/* End .ps-widget */}
+
+              {/* <div className="ps-widget bgc-white bdrs12 default-box-shadow2 p30 mb30 overflow-hidden position-relative">
+                <h4 className="title fz17 mb30">Que hay cerca?</h4>
+                <div className="row">
+                  <PropertyNearby />
+                </div>
+              </div> */}
+              {/* End .ps-widget */}
+
+              {/* <div className="ps-widget bgc-white bdrs12 default-box-shadow2 p30 mb30 overflow-hidden position-relative">
+                <h4 className="title fz17 mb30">Walkscore</h4>
+                <div className="row">
+                  <div className="col-md-12">
+                    <h4 className="fw400 mb20">
+                      10425 Tabor St Los Angeles CA 90034 USA
+                    </h4>
+                    <WalkScore />
+                  </div>
+                </div>
+              </div> */}
+              {/* End .ps-widget */}
+
+              <div className="ps-widget bgc-white bdrs12 default-box-shadow2 p30 mb30 overflow-hidden position-relative">
+                <h4 className="title fz17 mb30">Obtenen mas información</h4>
+                <InfoWithForm id={params.id} listings={listings}/>
+              </div>
+              {/* End .ps-widget */}
+            </div>
+            {/* End .col-8 */}
+
+            <div className="col-lg-4">
+              <div className="column">
+                <div className="default-box-shadow1 bdrs12 bdr1 p30 mb30-md bgc-white position-relative">
+                  <h4 className="form-title mb5">Programe una visita</h4>
+                  <p className="text">Elige tu dia preferido</p>
+                  <ScheduleTour id={params.id} listings={listings}/>
+                </div>
+                {/* End .Schedule a tour */}
+
+                <div className="agen-personal-info position-relative bgc-white default-box-shadow1 bdrs12 p30 mt30">
+                  <div className="widget-wrapper mb-0">
+                    <h6 className="title fz17 mb30">Obten mas información</h6>
+                    <ContactWithAgent id={params.id} listings={listings}/>
+                  </div>
+                </div>
+              </div>
+            </div>
+          </div>
+          {/* End .row */}
+
+          <div className="row mt30 align-items-center justify-content-between">
+            <div className="col-auto">
+              <div className="main-title">
+                <h2 className="title">Descubre mas propiedades</h2>
+                <p className="paragraph">
+                  Has tus sueños realidad con nosotros
+                </p>
+              </div>
+            </div>
+            {/* End header */}
+
+            <div className="col-auto mb30">
+              <div className="row align-items-center justify-content-center">
+                <div className="col-auto">
+                  <button className="featured-prev__active swiper_button">
+                    <i className="far fa-arrow-left-long" />
+                  </button>
+                </div>
+                {/* End prev */}
+
+                <div className="col-auto">
+                  <div className="pagination swiper--pagination featured-pagination__active" />
+                </div>
+                {/* End pagination */}
+
+                <div className="col-auto">
+                  <button className="featured-next__active swiper_button">
+                    <i className="far fa-arrow-right-long" />
+                  </button>
+                </div>
+                {/* End Next */}
+              </div>
+              {/* End .col for navigation and pagination */}
+            </div>
+            {/* End .col for navigation and pagination */}
+          </div>
+          {/* End .row */}
+
+          <div className="row">
+            <div className="col-lg-12">
+              <div className="property-city-slider">
+                <NearbySimilarProperty listings={listings} />
+              </div>
+            </div>
+          </div>
+          {/* End .row */}
+        </div>
+        {/* End .container */}
+      </section>
+      {/* End Property All Single V1  */}
+
+      {/* Start Our Footer */}
+      <section className="footer-style1 pt60 pb-0">
+        <Footer />
+      </section>
+      {/* End Our Footer */}
+    </>
+  );
+};
+
+export default SingleComponent;
