+++ src/components/property/property-single-style/single-v10/PropertyGallery.js
@@ -16,7 +16,7 @@ const images = [
 ];
 
 const PropertyGallery = ({ id }) => {
-  const data = listings.filter((elm) => elm.id == id)[0] || listings[0];
+  //const data = listings.filter((elm) => elm.id == id)[0] || listings[0];
   return (
     <>
       <Gallery>
